feat(std/node): consistent Node.js builtin shapes (#8274)
