+++ std/node/assert.ts
@@ -1,25 +1,42 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
-import {
-  assertEquals,
-  assertMatch,
-  assertNotEquals,
-  assertNotStrictEquals,
-  assertStrictEquals,
-  assertThrows,
-} from "../testing/asserts.ts";
-
 export { AssertionError } from "./assertion_error.ts";
-
-export {
-  assert,
-  assert as default,
-  assert as ok,
+import {
+  assertEquals as deepStrictEqual,
+  AssertionError,
+  assertMatch as match,
+  assertNotEquals as notDeepStrictEqual,
+  assertNotStrictEquals as notStrictEqual,
+  assertStrictEquals as strictEqual,
+  assertThrows as throws,
   fail,
 } from "../testing/asserts.ts";
 
-export const deepStrictEqual = assertEquals;
-export const notDeepStrictEqual = assertNotEquals;
-export const strictEqual = assertStrictEquals;
-export const notStrictEqual = assertNotStrictEquals;
-export const match = assertMatch;
-export const throws = assertThrows;
+function assert(expr: unknown, msg = ""): asserts expr {
+  if (!expr) {
+    throw new AssertionError(msg);
+  }
+}
+const ok = assert;
+export default assert;
+
+Object.assign(assert, {
+  deepStrictEqual,
+  fail,
+  match,
+  notDeepStrictEqual,
+  notStrictEqual,
+  ok,
+  strictEqual,
+  throws,
+});
+
+export {
+  deepStrictEqual,
+  fail,
+  match,
+  notDeepStrictEqual,
+  notStrictEqual,
+  ok,
+  strictEqual,
+  throws,
+};
