+++ std/node/crypto.ts
@@ -1,4 +1,4 @@
-import randomBytes from "./_crypto/randomBytes.ts";
-import { pbkdf2, pbkdf2Sync } from "./_crypto/pbkdf2.ts";
-
-export { pbkdf2, pbkdf2Sync, randomBytes };
+// Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
+export * from "./_crypto.ts";
+import * as m from "./_crypto.ts";
+export default m;
